# Sync optimizations