--- a/News-Android-App/src/main/java/de/luhmer/owncloudnewsreader/LoginDialogFragment.java
+++ b/News-Android-App/src/main/java/de/luhmer/owncloudnewsreader/LoginDialogFragment.java
@@ -55,6 +55,7 @@ import de.luhmer.owncloud.accountimporter.helper.OwnCloudAccount;
 import de.luhmer.owncloud.accountimporter.interfaces.IAccountImport;
 import de.luhmer.owncloudnewsreader.authentication.AuthenticatorActivity;
 import de.luhmer.owncloudnewsreader.database.DatabaseConnectionOrm;
+import de.luhmer.owncloudnewsreader.reader.HttpJsonRequest;
 import de.luhmer.owncloudnewsreader.reader.owncloud.OwnCloudReaderMethods;
 
 /**
@@ -316,6 +317,7 @@ public class LoginDialogFragment extends DialogFragment implements IAccountImpor
 		protected Integer doInBackground(Void... params) {
 
 			try {
+				HttpJsonRequest.getInstance().setCredentials(username, password);
 				String _version = OwnCloudReaderMethods.GetVersionNumber(getActivity(), oc_root_path);
 				if(_version != null)
 				{