… News-Android-App/src/main/java/de/luhmer/owncloudnewsreader/reader/IReader.java
… News-Android-App/src/main/java/de/luhmer/owncloudnewsreader/reader/OkHttpImageDownloader.java
… News-Android-App/src/main/java/de/luhmer/owncloudnewsreader/reader/owncloud/AsyncTask_GetApiVersion.java
… News-Android-App/src/main/java/de/luhmer/owncloudnewsreader/reader/owncloud/AsyncTask_GetFeeds.java
… News-Android-App/src/main/java/de/luhmer/owncloudnewsreader/reader/owncloud/AsyncTask_GetFolderTags.java
… News-Android-App/src/main/java/de/luhmer/owncloudnewsreader/reader/owncloud/AsyncTask_GetItems.java
… News-Android-App/src/main/java/de/luhmer/owncloudnewsreader/reader/owncloud/AsyncTask_GetOldItems.java
… News-Android-App/src/main/java/de/luhmer/owncloudnewsreader/reader/owncloud/AsyncTask_PerformItemStateChange.java
… News-Android-App/src/main/java/de/luhmer/owncloudnewsreader/reader/owncloud/OwnCloudConstants.java
… News-Android-App/src/main/java/de/luhmer/owncloudnewsreader/reader/owncloud/OwnCloudReaderMethods.java
… News-Android-App/src/main/java/de/luhmer/owncloudnewsreader/reader/owncloud/OwnCloud_Reader.java
… News-Android-App/src/main/java/de/luhmer/owncloudnewsreader/reader/owncloud/apiv1/APIv1.java
… News-Android-App/src/main/java/de/luhmer/owncloudnewsreader/reader/owncloud/apiv2/APIv2.java
… News-Android-App/src/main/java/de/luhmer/owncloudnewsreader/services/OwnCloudSyncService.java
… News-Android-App/src/main/java/de/luhmer/owncloudnewsreader/services/SyncItemStateService.java