This is not ready to be merged yet, I wanted to get some feedback first for those changes before I clean it up.

These are some experiments to improve the time needed for synchronizations:

* Use [OkHttp](http://square.github.io/okhttp/) instead of HTTPUrlConnection
  * Also use OkHttp to download images
  * This also simplifies the network code a bit
* Start AsyncTasks in parallel
  * Synchronize the completion of all tasks using a [CountDownLatch](https://developer.android.com/reference/java/util/concurrent/CountDownLatch.html)
* Only get the api version on the first sync

A quick benchmark shows an improvement on average from 4,68 s to 3,20 s needed for synchronization:
```
origin/master
-------------
V/OwnCloudSyncService﹕ Time needed (sync) 00:00:04.906
V/OwnCloudSyncService﹕ Time needed (sync) 00:00:04.624
V/OwnCloudSyncService﹕ Time needed (sync) 00:00:04.505
V/OwnCloudSyncService﹕ Time needed (sync) 00:00:04.703
V/OwnCloudSyncService﹕ Time needed (sync) 00:00:04.679

Average: 4,68 s

sync-optimizations
----------
V/OwnCloudSyncService﹕ Time needed (sync): 00:00:03.288
V/OwnCloudSyncService﹕ Time needed (sync): 00:00:03.305
V/OwnCloudSyncService﹕ Time needed (sync): 00:00:03.128
V/OwnCloudSyncService﹕ Time needed (sync): 00:00:03.134
V/OwnCloudSyncService﹕ Time needed (sync): 00:00:03.156

Average: 3,20 s
```