@David-Development 2015-09-20
Removing IReader interface: Great, this was a fragment from the first implementation of multiple api endpoints.. Not needed any longer. Thanks

One suggestion to probably make the sync even faster:
Encapsulate every execute call in OwnCloud_Reader AsyncTask by AsyncTaskHelper.StartAsyncTask(new [AsyncTask...]);
This could increase the performance as well. Because they can be executed in parallel.

What are the advantages of using the okhttp lib? (haven't used it before)