@schaal 2015-09-20
I feel it has a nicer API and it has supposedly better performance. It also supports connection pooling, which reduces the latency when connecting multiple times to the same host.

I'll try using the AsyncTaskHelper and check whether it improves performance further.