@schaal 2015-09-26
I now cleaned up the patch set for you to review. I didn't include the parallelization of the AsyncTasks as it doesn't seem to improve things much and makes it more fragile.

Updated benchmark:
```
origin/master @ 7c646e3e
-------------
de.luhmer.owncloudnewsreader D/OwnCloudSyncService﹕ Time needed (sync): 00:00:04.540
de.luhmer.owncloudnewsreader D/OwnCloudSyncService﹕ Time needed (sync): 00:00:04.480
de.luhmer.owncloudnewsreader D/OwnCloudSyncService﹕ Time needed (sync): 00:00:04.489
de.luhmer.owncloudnewsreader D/OwnCloudSyncService﹕ Time needed (sync): 00:00:04.420
de.luhmer.owncloudnewsreader D/OwnCloudSyncService﹕ Time needed (sync): 00:00:04.420

Average: 4,47 s

sync-optimizations
----------
de.luhmer.owncloudnewsreader D/OwnCloudSyncService﹕ Time needed (sync): 00:00:03.249
de.luhmer.owncloudnewsreader D/OwnCloudSyncService﹕ Time needed (sync): 00:00:03.100
de.luhmer.owncloudnewsreader D/OwnCloudSyncService﹕ Time needed (sync): 00:00:03.099
de.luhmer.owncloudnewsreader D/OwnCloudSyncService﹕ Time needed (sync): 00:00:03.100
de.luhmer.owncloudnewsreader D/OwnCloudSyncService﹕ Time needed (sync): 00:00:03.090

Average: 3,13 s
```